requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
astroid==2.4.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coloredlogs==14.0
config==0.5.0.post0
cryptography==2.9.2
decorator==4.4.2
discord.py==1.3.3
distlib==0.3.0
filelock==3.0.12
future==0.18.2
humanfriendly==8.2
idna==2.9
isort==4.3.21
lavalink==3.1.0
lazy-object-proxy==1.4.3
lyricwikia==0.1.11
mccabe==0.6.1
more-itertools==8.3.0
multidict==4.7.6
mutagen==1.44.0
numpy==1.18.4
packaging==20.3
pathlib==1.0.1
pipenv==2018.11.26
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pylint==2.5.2
PyNaCl==1.3.0
pyparsing==2.4.7
pyreadline==2.1
pyTelegramBotAPI==0.3.0
pytest==5.4.2
python-telegram-bot==12.7
pytube3==9.6.4
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
soupsieve==2.0.1
spotdl==2.0.5
spotipy==2.12.0
toml==0.10.1
tornado==6.0.4
tqdm==4.46.0
typing-extensions==3.7.4.2
unicode-slugify==0.1.3
Unidecode==1.1.1
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
wcwidth==0.1.9
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
youtube-dl==2020.5.8
